+++ server/src/main/java/org/opensearch/index/store/remote/file/OnDemandBlockSnapshotIndexInput.java
@@ -8,8 +8,6 @@
 
 package org.opensearch.index.store.remote.file;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.apache.lucene.store.FSDirectory;
 import org.apache.lucene.store.IndexInput;
 import org.opensearch.index.snapshots.blobstore.BlobStoreIndexShardSnapshot.FileInfo;
@@ -26,8 +24,6 @@ import java.io.IOException;
  * @opensearch.internal
  */
 public class OnDemandBlockSnapshotIndexInput extends OnDemandBlockIndexInput {
-    private static final Logger logger = LogManager.getLogger(OnDemandBlockSnapshotIndexInput.class);
-
     /**
      * Where this class fetches IndexInput parts from
      */
@@ -48,7 +44,7 @@ public class OnDemandBlockSnapshotIndexInput extends OnDemandBlockIndexInput {
     protected final String fileName;
 
     /**
-     * part size  in bytes
+     * Maximum size in bytes of snapshot file parts.
      */
     protected final long partSize;
 
@@ -104,7 +100,15 @@ public class OnDemandBlockSnapshotIndexInput extends OnDemandBlockIndexInput {
         super(builder);
         this.transferManager = transferManager;
         this.fileInfo = fileInfo;
-        this.partSize = fileInfo.partSize().getBytes();
+        if (fileInfo.partSize() != null) {
+            this.partSize = fileInfo.partSize().getBytes();
+        } else {
+            // Repository implementations can define a size at which to split files
+            // into multiple objects in the repository. If partSize() is null, then
+            // no splitting happens, so default to Long.MAX_VALUE here to have the
+            // same effect. See {@code BlobStoreRepository#chunkSize()}.
+            this.partSize = Long.MAX_VALUE;
+        }
         this.fileName = fileInfo.physicalName();
         this.directory = directory;
         this.originalFileSize = fileInfo.length();
@@ -131,6 +135,10 @@ public class OnDemandBlockSnapshotIndexInput extends OnDemandBlockIndexInput {
 
         final long blockStart = getBlockStart(blockId);
         final long blockEnd = blockStart + getActualBlockSize(blockId);
+
+        // If the snapshot file is chunked, we must account for this by
+        // choosing the appropriate file part and updating the position
+        // accordingly.
         final int part = (int) (blockStart / partSize);
         final long partStart = part * partSize;